Prepare structure for data and rename SyncRulesWizard to MetadataSyncRulesWizard: - Move current wizard and steps under metadata folder
