+++ src/components/app/Root.jsx
@@ -9,7 +9,7 @@ import MetadataPage from "../synchronization-page/MetadataPage";
 import DataPage from "../synchronization-page/DataPage";
 import DeletedObjectsPage from "../synchronization-page/DeletedObjectsPage";
 import HistoryPage from "../history-list-page/HistoryPage";
-import SyncRulesWizard from "../rules-creation-page/SyncRulesWizard";
+import MetadadaSyncRulesWizard from "../rules-creation-page/metadata/MetadataSyncRulesWizard";
 import MetadataSyncRulesPage from "../rules-list-page/MetadataSyncRulesPage";
 import DataSyncRulesPage from "../rules-list-page/DataSyncRulesPage";
 
@@ -53,7 +53,7 @@ class Root extends React.Component {
 
                 <Route
                     path={"/metadata-synchronization-rules/:action(new|edit)/:id?"}
-                    render={props => <SyncRulesWizard {...this.props} {...props} />}
+                    render={props => <MetadadaSyncRulesWizard {...this.props} {...props} />}
                 />
 
                 <Route
