+++ src/components/rules-creation-page/metadata/steps/GeneralInfoStep.jsx
